Flairs: Refactor to new FlairsSettings component
It was getting messy in the Settings component, so moving all that logic to a new subcomponent
